requirements.txt
openai==0.28.1
tiktoken==0.6.0
langchain==0.1.20
chromadb==0.5.0
faiss-cpu
streamlit
streamlit_option_menu
streamlit_extras
plotly
openpyxl
scipy
scikit-learn